requirements.txt
scrapy>=2.11.0
PyYAML>=6.0.1
beautifulsoup4>=4.12.3
markdownify>=0.12.1
flask>=2.3
pymilvus>=2.4.0
transformers>=4.38
torch>=2.1
requests>=2.31
# Core
pymilvus>=2.4.0

# LangChain (documentos, splitters, embeddings)
langchain>=0.1.0
langchain-community>=0.0.30

# Embeddings
sentence-transformers>=2.6.0

# HTML parsing
beautifulsoup4>=4.12.0

# OCR
pillow>=10.0.0
pytesseract>=0.3.10